% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_directed_louvain.R
\name{cluster_directed_louvain}
\alias{cluster_directed_louvain}
\title{Directed Louvain Community Detection}
\usage{
cluster_directed_louvain(graph)
}
\description{
Directed Louvain Community Detection
}
