% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_node_closeness.R
\name{cluster_node_closeness}
\alias{cluster_node_closeness}
\title{Community structure detection based on node closeness centrality}
\usage{
cluster_node_closeness(graph)
}
\arguments{
\item{graph}{The graph to analyze}
}
\description{
This algorithm detects communities by calculating the closeness centrality measures of nodes and assigns communities based on the subsequent subgraphs. For more information see Details.
}
\details{
The algorithm first calculates closeness centrality for all nodes in the graph and then starts from the node with the largest closeness centrality extracts the sub-graph set it as its "cluster". The remaining nodes are assigned to a separate cluster. This is preformed iteratively until all possible communities are identified.

In the case of an overlap of nodes, the node with higher closeness centrality gets it. As a result, there are is no overlap in nodes between clusters. For choice of number of communities, the number which maximizes modularity is chosen.
}
\examples{
library(igraphdata)
data("karate")
nc <- cluster_node_closeness(karate)
plot(
nc,
karate,
main= "Node Closeness Clustering"
)

nc
}
