% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_node_degree.R
\name{cluster_node_degree}
\alias{cluster_node_degree}
\title{Community structure detection based on node degree centrality}
\usage{
cluster_node_degree(graph)
}
\arguments{
\item{graph}{The graph to analyze}
}
\description{
This algorithm detects communities by calculating the degree centrality measures of nodes and assigns communities based on the subsequent subgraphs. The steps outlined are the same as those outlined in `cluster_node_closeness()` with degree centrality being used instead of closeness.
}
\examples{
library(igraphdata)
data("karate")
nd <- cluster_node_degree(karate)
plot(
nd,
karate,
main= "Node Degree Clustering"
)

nd
}
