% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.degree.betweenness-package.R
\docType{package}
\name{ig.degree.betweenness-package}
\alias{ig.degree.betweenness}
\alias{ig.degree.betweenness-package}
\title{ig.degree.betweenness: "Node+Edge Betweenness Community Detection Algorithm for 'igraph' Objects"}
\description{
Implements the "Node + Edge Betweenness" community detection algorithm for network analysis using 'igraph' objects. This algorithm combines node degree and betweenness centrality measures to identify communities within networks, with a gradient evident in social partitioning. The package provides functions for implementation of the algorithm, visualization, and analysis of the resulting community structure along with the original dataset used in the publication. Methods are based on results from Smith, Pittman and Xu (2024) \doi{10.48550/arXiv.2411.01394}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benyamindsmith/ig.degree.betweenness}
  \item Report bugs at \url{https://github.com/benyamindsmith/ig.degree.betweenness/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Smith \email{benyamin.smith@mail.utoronto.ca} (\href{https://orcid.org/0009-0007-2206-0177}{ORCID})

Authors:
\itemize{
  \item Tyler Pittman \email{Tyler.Pittman@uhn.ca} (\href{https://orcid.org/0000-0002-5013-6980}{ORCID})
  \item Wei Xu \email{Wei.Xu@uhn.ca} (\href{https://orcid.org/0000-0002-0257-8856}{ORCID})
}

}
\keyword{internal}
