% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd_graphs.R
\name{lcd_graph}
\alias{lcd_graph}
\title{Linearized Chord Diagram Graph Generating functions}
\usage{
lcd_graph(n, m = 1, directed = TRUE)
}
\description{
Linearized Chord Diagram Graph Generating functions
}
