% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_degree_betweenness.R
\name{cluster_degree_betweenness}
\alias{cluster_degree_betweenness}
\title{Community structure detection based on node degree centrality and edge betweeeness}
\usage{
cluster_degree_betweenness(graph)
}
\arguments{
\item{graph}{The graph to analyze}
}
\description{
This algorithm detects communities by calculating the degree centrality measures of nodes and edge betweenness.
}
\details{
This can be thought of as an alternative version of `igraph::cluster_edge_betweeness()`.

The function iteratively removes edges based on their betweenness centrality and the degree of their adjacent nodes. At each iteration, it identifies the edge with the highest betweenness centrality among those connected to nodes with the highest degree.It then removes that edge and recalculates the modularity of the resulting graph. The process continues until all edges have been assessed or until no further subgraph can be created.

The optimal number of communites are chosen based on maximization of modularity
}
\examples{
library(igraphdata)
data("karate")
ndb <- cluster_degree_betweenness(karate)
plot(
ndb,
karate,
main= "Node Degree Clustering"
)

ndb
}
