% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simplified_edgeplot.R
\name{plot_simplified_edgeplot}
\alias{plot_simplified_edgeplot}
\title{Plot Simplified Edgeplot
This function generates a simplified edge plot of an igraph object, optionally highlighting communities if provided.}
\usage{
plot_simplified_edgeplot(graph, communities = NULL, ...)
}
\arguments{
\item{graph}{igraph object}

\item{communities}{optional; A communities object}

\item{...}{other arguments to be passed to the `plot()` function}
}
\description{
Plot Simplified Edgeplot
This function generates a simplified edge plot of an igraph object, optionally highlighting communities if provided.
}
\examples{
# Create an igraph object
my_graph <- igraph::make_ring(10)
# Plot the simplified edge plot
plot_simplified_edgeplot(my_graph)

}
