% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_unlabeled_graph.R
\name{prep_unlabeled_graph}
\alias{prep_unlabeled_graph}
\title{Prepared Unlabeled Graph to work with Degree-Betweenness Algorithm}
\usage{
prep_unlabeled_graph(graph)
}
\description{
Prepared Unlabeled Graph to work with Degree-Betweenness Algorithm
}
\examples{
library(igraph)
library(igraphdata)
library(ig.degree.betweenness)
data("UKfaculty")
# Making graph undirected so it looks nicer when its plotted
uk_faculty <- prep_unlabeled_graph(UKfaculty) |>
  as.undirected()

ndb <- cluster_degree_betweenness(uk_faculty)

plot(
ndb,
uk_faculty,
main= "Node Degree Clustering"
)

ndb
}
